{"schema":"decodex/commit/1","summary":"Split Decodex app telemetry widgets","authority":"XY-1115"}
