+++ apps/decodex-app/Sources/DecodexApp/AccountPanelView.swift
@@ -677,7 +677,7 @@ struct AccountRowView: View {
 	}
 }
 
-private enum AccountPanelLayout {
+enum AccountPanelLayout {
 	static let accountListScrollSpace = "account-list-scroll"
 	static let screenVerticalMargin: CGFloat = 44
 	static let panelVerticalPadding: CGFloat = 18
@@ -746,898 +746,6 @@ private struct AccountListScrollIndicatorView: View {
 		let progress = min(1, max(0, scrollOffset / maxScrollOffset))
 
 		return maxThumbOffset * progress
-	}
-}
-
-private struct AccountProfileAggregate: Equatable {
-	let accountCount: Int
-	let lifetimeTokens: Int?
-	let peakDailyTokens: Int?
-	let longestTaskSeconds: Int?
-	let currentStreakDays: Int?
-	let longestStreakDays: Int?
-	let dailyUsage: [AccountProfileDailyUsage]
-
-	static func make(accounts: [CodexAccount]) -> AccountProfileAggregate? {
-		var lifetimeTokens: Int?
-		var peakFallbackTokens: Int?
-		var longestTaskSeconds: Int?
-		var currentStreakDays: Int?
-		var longestStreakDays: Int?
-		var usageByDate: [String: Int] = [:]
-
-		for account in accounts {
-			if let value = account.profileLifetimeTokens {
-				lifetimeTokens = (lifetimeTokens ?? 0) + value
-			}
-			if let value = account.profilePeakDailyTokens {
-				peakFallbackTokens = (peakFallbackTokens ?? 0) + value
-			}
-			if let value = account.profileLongestTaskSeconds {
-				longestTaskSeconds = max(longestTaskSeconds ?? 0, value)
-			}
-			if let value = account.profileCurrentStreakDays {
-				currentStreakDays = max(currentStreakDays ?? 0, value)
-			}
-			if let value = account.profileLongestStreakDays {
-				longestStreakDays = max(longestStreakDays ?? 0, value)
-			}
-			for record in account.recentProfileDailyUsage {
-				usageByDate[record.date, default: 0] += record.tokens
-			}
-		}
-
-		let dailyUsage = usageByDate
-			.map { AccountProfileDailyUsage(date: $0.key, tokens: $0.value) }
-			.sorted { $0.date < $1.date }
-		let peakDailyTokens = dailyUsage.map(\.tokens).max() ?? peakFallbackTokens
-		let aggregate = AccountProfileAggregate(
-			accountCount: accounts.count,
-			lifetimeTokens: lifetimeTokens,
-			peakDailyTokens: peakDailyTokens,
-			longestTaskSeconds: longestTaskSeconds,
-			currentStreakDays: currentStreakDays,
-			longestStreakDays: longestStreakDays,
-			dailyUsage: dailyUsage
-		)
-
-		return aggregate.hasProfileSummary ? aggregate : nil
-	}
-
-	var hasProfileSummary: Bool {
-		lifetimeTokens != nil
-			|| peakDailyTokens != nil
-			|| longestTaskSeconds != nil
-			|| currentStreakDays != nil
-			|| longestStreakDays != nil
-			|| dailyUsage.isEmpty == false
-	}
-}
-
-private struct AccountTelemetryMatrixView: View {
-	let aggregate: AccountProfileAggregate?
-	let usageEstimate: AccountUsageEstimate?
-	let accounts: [CodexAccount]
-	@Environment(\.colorScheme) private var colorScheme
-
-	var body: some View {
-		VStack(alignment: .leading, spacing: AccountPanelLayout.telemetryRowSpacing) {
-			if let aggregate {
-				AccountProfileOverviewView(aggregate: aggregate)
-			}
-
-			if let usageEstimate {
-				AccountPoolUsageEstimateView(estimate: usageEstimate, accounts: accounts)
-			}
-		}
-		.padding(.horizontal, AccountPanelLayout.telemetryHorizontalPadding)
-		.padding(.top, AccountPanelLayout.telemetryTopPadding)
-		.padding(.bottom, AccountPanelLayout.telemetryBottomPadding)
-		.frame(maxWidth: .infinity, alignment: .leading)
-		.background {
-			RoundedRectangle(cornerRadius: 9, style: .continuous)
-				.fill(surfaceFill)
-		}
-		.clipShape(RoundedRectangle(cornerRadius: 9, style: .continuous))
-		.id(colorScheme == .dark ? "telemetry-matrix-dark" : "telemetry-matrix-light")
-	}
-
-	private var surfaceFill: Color {
-		colorScheme == .dark
-			? Color(red: 0.08, green: 0.095, blue: 0.13).opacity(0.34)
-			: Color(red: 0.9, green: 0.94, blue: 0.98).opacity(0.48)
-	}
-}
-
-private struct AccountProfileOverviewView: View {
-	let aggregate: AccountProfileAggregate
-	@Environment(\.colorScheme) private var colorScheme
-
-	var body: some View {
-		VStack(alignment: .leading, spacing: 5) {
-			HStack(alignment: .firstTextBaseline, spacing: 5) {
-				PanelMetricIconView(
-					symbol: "sum",
-					tint: PanelPalette.usageCyan(colorScheme).opacity(0.9)
-				)
-
-				Text("All accounts")
-					.font(PanelFont.metricLabel)
-					.foregroundStyle(PanelPalette.secondaryText(colorScheme))
-					.lineLimit(1)
-
-				Spacer(minLength: 6)
-
-				Text("\(aggregate.accountCount) accounts")
-					.font(PanelFont.tertiary)
-					.foregroundStyle(PanelPalette.secondaryText(colorScheme).opacity(0.72))
-					.lineLimit(1)
-			}
-
-			HStack(spacing: 5) {
-				ForEach(Array(metrics.enumerated()), id: \.offset) { index, metric in
-					HStack(alignment: .firstTextBaseline, spacing: 3) {
-						Text(metric.label)
-							.font(PanelFont.usageLabel)
-							.foregroundStyle(PanelPalette.secondaryText(colorScheme).opacity(0.82))
-							.lineLimit(1)
-
-						Text(metric.value)
-							.font(PanelFont.usageValue)
-							.foregroundStyle(index == 0 ? primaryMetricColor : PanelPalette.secondaryText(colorScheme))
-							.monospacedDigit()
-							.lineLimit(1)
-							.minimumScaleFactor(0.72)
-					}
-
-					if index < metrics.count - 1 {
-						Spacer(minLength: 3)
-					}
-				}
-			}
-			.frame(height: 16)
-
-			if aggregate.dailyUsage.isEmpty == false {
-				AccountProfileDailyUsageStripView(records: aggregate.dailyUsage)
-			}
-		}
-		.frame(maxWidth: .infinity, alignment: .leading)
-		.accessibilityLabel(accessibilityLabel)
-	}
-
-	private var metrics: [(label: String, value: String)] {
-		[
-			aggregate.lifetimeTokens.map { ("tok", formatCompactCount($0)) },
-			aggregate.peakDailyTokens.map { ("peak", formatCompactCount($0)) },
-			streakText.map { ("streak", $0) },
-			aggregate.longestTaskSeconds
-				.flatMap(formatActivityDuration)
-				.map { ("task", $0) },
-		]
-		.compactMap { $0 }
-	}
-
-	private var streakText: String? {
-		if let current = aggregate.currentStreakDays,
-			let longest = aggregate.longestStreakDays
-		{
-			return "\(current)/\(longest)d"
-		}
-		if let current = aggregate.currentStreakDays {
-			return "\(current)d"
-		}
-		if let longest = aggregate.longestStreakDays {
-			return "\(longest)d"
-		}
-
-		return nil
-	}
-
-	private var primaryMetricColor: Color {
-		PanelPalette.primaryText(colorScheme).opacity(colorScheme == .dark ? 0.92 : 0.86)
-	}
-
-	private var accessibilityLabel: String {
-		"All account profile totals, " + metrics.map { "\($0.label) \($0.value)" }.joined(separator: ", ")
-	}
-}
-
-struct AccountPoolUsageEstimateView: View {
-	let estimate: AccountUsageEstimate
-	let accounts: [CodexAccount]
-	@Environment(\.colorScheme) private var colorScheme
-
-	var body: some View {
-		VStack(alignment: .leading, spacing: 3) {
-			HStack(spacing: 5) {
-				ForEach(Array(metrics.enumerated()), id: \.offset) { index, metric in
-					AccountPoolUsageMetricView(
-						title: metric.title,
-						value: metric.value,
-						tint: metric.tint
-					)
-
-					if index < metrics.count - 1 {
-						Spacer(minLength: 3)
-					}
-				}
-			}
-			.frame(height: 16)
-
-			if estimate.accountEstimateCount < estimate.accountCount {
-				Text("\(estimate.accountEstimateCount)/\(estimate.accountCount) accounts measured")
-					.font(PanelFont.tertiary)
-					.foregroundStyle(PanelPalette.secondaryText(colorScheme).opacity(0.72))
-					.lineLimit(1)
-			}
-		}
-		.frame(maxWidth: .infinity, alignment: .leading)
-		.accessibilityLabel(accessibilityLabel)
-	}
-
-	private var metrics: [(title: String, value: String, tint: Color)] {
-		[
-			(
-				"Pool used",
-				formatUsagePercent(estimate.totalUsedOfCapacityPercent),
-				poolUsageTint
-			),
-			("Day Δ", dayDeltaText, dayDeltaTint),
-			(
-				"Daily avg",
-				formatDailyUsageRate(estimate.averageDailyPoolPercent),
-				PanelPalette.secondaryText(colorScheme)
-			),
-		]
-	}
-
-	private var accessibilityLabel: String {
-		"Pool usage over \(estimate.windowDays) days: \(formatUsagePercent(estimate.totalUsedOfCapacityPercent)) used, daily change \(dayDeltaText), average \(formatUsagePercent(estimate.averageDailyPoolPercent)) per day"
-	}
-
-	private var dayDeltaText: String {
-		guard let delta = dayDeltaPercentagePoints else {
-			return "-"
-		}
-
-		return formatPercentagePointDelta(delta)
-	}
-
-	private var poolUsageTint: Color {
-		let used = estimate.totalUsedOfCapacityPercent
-		if used >= 90 {
-			return PanelPalette.destructive(colorScheme)
-		}
-		if used >= 75 {
-			return PanelPalette.warning(colorScheme)
-		}
-
-		return PanelPalette.routeAccent(colorScheme)
-	}
-
-	private var dayDeltaTint: Color {
-		guard let delta = dayDeltaPercentagePoints else {
-			return PanelPalette.secondaryText(colorScheme)
-		}
-		if delta > 0.05 {
-			if estimate.totalUsedOfCapacityPercent >= 90 {
-				return PanelPalette.destructive(colorScheme)
-			}
-			if estimate.totalUsedOfCapacityPercent >= 75 {
-				return PanelPalette.warning(colorScheme)
-			}
-
-			return PanelPalette.capacityAccent(colorScheme)
-		}
-		if delta < -0.05 {
-			return PanelPalette.secondaryText(colorScheme)
-		}
-
-		return PanelPalette.secondaryText(colorScheme)
-	}
-
-	private var dayDeltaPercentagePoints: Double? {
-		let measuredAccounts = accounts.filter { account in
-			account.sevenDayUsedPercent != nil
-		}
-		guard measuredAccounts.isEmpty == false, estimate.totalCapacityPercent > 0 else {
-			return nil
-		}
-
-		let latestDate = measuredAccounts
-			.flatMap(\.recentUsageRecords)
-			.map(\.date)
-			.max()
-		guard let latestDate else {
-			return estimate.averageDailyPoolPercent
-		}
-		guard let previousDate = previousUsageDate(before: latestDate) else {
-			return estimate.averageDailyPoolPercent
-		}
-
-		let previousRecords = measuredAccounts.compactMap { account in
-			usageRecord(for: account, on: previousDate).map { (account, $0) }
-		}
-		guard previousRecords.count == measuredAccounts.count else {
-			return estimate.averageDailyPoolPercent
-		}
-		let previousUsedPercent = previousRecords.reduce(0) { total, pair in
-			let (account, record) = pair
-
-			return total + record.usedPercent * (record.capacityMultiplier ?? account.capacityWeight)
-		}
-		let previousPoolPercent =
-			(Double(previousUsedPercent) / Double(estimate.totalCapacityPercent)) * 100
-
-		return estimate.totalUsedOfCapacityPercent - previousPoolPercent
-	}
-
-	private func usageRecord(
-		for account: CodexAccount,
-		on date: String
-	) -> AccountUsageRecord? {
-		account.recentUsageRecords
-			.filter { record in record.date == date }
-			.max { left, right in
-				left.checkedAtUnixEpoch < right.checkedAtUnixEpoch
-			}
-	}
-
-	private func previousUsageDate(before value: String) -> String? {
-		let formatter = DateFormatter()
-		formatter.locale = Locale(identifier: "en_US_POSIX")
-		formatter.dateFormat = "yyyy-MM-dd"
-		let calendar = Calendar(identifier: .gregorian)
-
-		guard let date = formatter.date(from: value),
-			let previousDate = calendar.date(byAdding: .day, value: -1, to: date)
-		else {
-			return nil
-		}
-
-		return formatter.string(from: previousDate)
-	}
-}
-
-struct AccountPoolUsageMetricView: View {
-	let title: String
-	let value: String
-	let tint: Color
-	@Environment(\.colorScheme) private var colorScheme
-
-	var body: some View {
-		HStack(alignment: .firstTextBaseline, spacing: 3) {
-			Text(title)
-				.font(PanelFont.usageLabel)
-				.foregroundStyle(PanelPalette.secondaryText(colorScheme).opacity(0.82))
-				.lineLimit(1)
-
-			Text(value)
-				.font(PanelFont.usageValue)
-				.foregroundStyle(tint.opacity(colorScheme == .dark ? 0.94 : 0.78))
-				.monospacedDigit()
-				.lineLimit(1)
-				.minimumScaleFactor(0.72)
-		}
-		.lineLimit(1)
-	}
-}
-
-struct AccountUsageSummaryView: View {
-	let account: CodexAccount
-	let currentTime: Date
-
-	var body: some View {
-		VStack(spacing: 5) {
-			if account.hasProfileSummary {
-				AccountProfileSummaryView(account: account)
-			}
-
-			if account.hasPrimaryUsageData {
-				AccountUsageMeterView(
-					label: account.windowLabel(seconds: account.primaryWindowSeconds),
-					remainingPercent: account.primaryRemainingPercent,
-					resetAtUnixEpoch: account.primaryResetsAtUnixEpoch,
-					dailyAveragePercent: account.sevenDayAveragePercent(
-						forWindowSeconds: account.primaryWindowSeconds
-					),
-					tone: account.usageTone(remainingPercent: account.primaryRemainingPercent),
-					currentTime: currentTime
-				)
-			}
-
-			if account.hasSecondaryUsageData {
-				AccountUsageMeterView(
-					label: account.windowLabel(seconds: account.secondaryWindowSeconds),
-					remainingPercent: account.secondaryRemainingPercent,
-					resetAtUnixEpoch: account.secondaryResetsAtUnixEpoch,
-					dailyAveragePercent: account.sevenDayAveragePercent(
-						forWindowSeconds: account.secondaryWindowSeconds
-					),
-					tone: account.usageTone(remainingPercent: account.secondaryRemainingPercent),
-					currentTime: currentTime
-				)
-			}
-		}
-		.frame(maxWidth: .infinity)
-		.padding(.horizontal, 1)
-		.padding(.vertical, 1)
-	}
-}
-
-struct AccountProfileSummaryView: View {
-	let account: CodexAccount
-	@Environment(\.colorScheme) private var colorScheme
-
-	var body: some View {
-		VStack(spacing: 4) {
-			if metrics.isEmpty == false {
-				HStack(alignment: .firstTextBaseline, spacing: 5) {
-					PanelMetricIconView(
-						symbol: "chart.bar.xaxis",
-						tint: PanelPalette.secondaryText(colorScheme).opacity(0.82)
-					)
-
-					ForEach(Array(metrics.enumerated()), id: \.offset) { index, metric in
-						HStack(alignment: .firstTextBaseline, spacing: 3) {
-							Text(metric.label)
-								.font(PanelFont.usageLabel)
-								.foregroundStyle(PanelPalette.secondaryText(colorScheme).opacity(0.82))
-								.lineLimit(1)
-
-							Text(metric.value)
-								.font(PanelFont.usageValue)
-								.foregroundStyle(valueColor(index: index))
-								.monospacedDigit()
-								.lineLimit(1)
-						}
-
-						if index < metrics.count - 1 {
-							Spacer(minLength: 3)
-						}
-					}
-				}
-				.frame(height: 16)
-			}
-
-			if account.recentProfileDailyUsage.isEmpty == false {
-				AccountProfileDailyUsageStripView(records: account.recentProfileDailyUsage)
-			}
-		}
-		.accessibilityLabel(accessibilityLabel)
-	}
-
-	private var metrics: [(label: String, value: String)] {
-		[
-			account.profileLifetimeTokens.map { ("tok", formatCompactCount($0)) },
-			account.profilePeakDailyTokensForDisplay.map { ("peak", formatCompactCount($0)) },
-			streakText.map { ("streak", $0) },
-			account.profileLongestTaskSeconds
-				.flatMap(formatActivityDuration)
-				.map { ("task", $0) },
-		]
-		.compactMap { $0 }
-	}
-
-	private var streakText: String? {
-		if let current = account.profileCurrentStreakDays,
-			let longest = account.profileLongestStreakDays
-		{
-			return "\(current)/\(longest)d"
-		}
-		if let current = account.profileCurrentStreakDays {
-			return "\(current)d"
-		}
-		if let longest = account.profileLongestStreakDays {
-			return "\(longest)d"
-		}
-
-		return nil
-	}
-
-	private var accessibilityLabel: String {
-		metrics.map { "\($0.label) \($0.value)" }.joined(separator: ", ")
-	}
-
-	private func valueColor(index: Int) -> Color {
-		index == 0
-			? PanelPalette.primaryText(colorScheme).opacity(colorScheme == .dark ? 0.92 : 0.86)
-			: PanelPalette.secondaryText(colorScheme)
-	}
-}
-
-struct AccountProfileDailyUsageStripView: View {
-	let records: [AccountProfileDailyUsage]
-	@Environment(\.colorScheme) private var colorScheme
-
-	var body: some View {
-		HStack(spacing: 2) {
-			ForEach(Array(displayRecords.enumerated()), id: \.offset) { _, record in
-				RoundedRectangle(cornerRadius: 2, style: .continuous)
-					.fill(tileColor(tokens: record.tokens))
-					.frame(width: 6, height: 9)
-					.help("\(compactUsageDate(record.date)): \(formatCompactCount(record.tokens)) tokens")
-			}
-		}
-		.frame(maxWidth: .infinity, alignment: .leading)
-		.frame(height: 11)
-		.accessibilityHidden(true)
-	}
-
-	private var displayRecords: [AccountProfileDailyUsage] {
-		Array(records.sorted { $0.date < $1.date }.suffix(36))
-	}
-
-	private var peakTokens: Int {
-		max(1, displayRecords.map(\.tokens).max() ?? 1)
-	}
-
-	private func tileColor(tokens: Int) -> Color {
-		let intensity = max(0.16, min(1, Double(tokens) / Double(peakTokens)))
-		return PanelPalette.usageCyan(colorScheme).opacity(0.24 + 0.62 * intensity)
-	}
-}
-
-struct AccountUsageMeterView: View {
-	let label: String
-	let remainingPercent: Int?
-	let resetAtUnixEpoch: Int?
-	let dailyAveragePercent: Double?
-	let tone: AccountTone
-	let currentTime: Date
-	@Environment(\.colorScheme) private var colorScheme
-
-	var body: some View {
-		VStack(alignment: .leading, spacing: 3) {
-			HStack(spacing: 5) {
-				Text(label)
-					.font(PanelFont.usageLabel)
-					.frame(width: 28, alignment: .leading)
-					.foregroundStyle(PanelPalette.secondaryText(colorScheme))
-
-				Text(remainingText)
-					.font(PanelFont.usageValue)
-					.frame(width: 62, alignment: .leading)
-					.foregroundStyle(valueColor)
-					.monospacedDigit()
-
-				if let dailyAverageText {
-					HStack(alignment: .firstTextBaseline, spacing: 3) {
-						Text("avg")
-							.font(PanelFont.usageLabel)
-							.foregroundStyle(PanelPalette.secondaryText(colorScheme).opacity(0.82))
-							.lineLimit(1)
-
-						Text(dailyAverageText)
-							.font(PanelFont.usageValue)
-							.foregroundStyle(PanelPalette.secondaryText(colorScheme))
-							.monospacedDigit()
-							.lineLimit(1)
-							.minimumScaleFactor(0.78)
-					}
-					.layoutPriority(1)
-				}
-
-				Spacer(minLength: 2)
-
-				Text(resetDisplay.short)
-					.font(PanelFont.usageValue)
-					.foregroundStyle(PanelPalette.secondaryText(colorScheme).opacity(colorScheme == .dark ? 0.82 : 0.9))
-					.monospacedDigit()
-					.lineLimit(1)
-
-				if resetDisplay.date.isEmpty == false {
-					Text(resetDisplay.date)
-						.font(PanelFont.tertiary)
-						.foregroundStyle(PanelPalette.secondaryText(colorScheme).opacity(colorScheme == .dark ? 0.68 : 0.78))
-						.lineLimit(1)
-						.truncationMode(.middle)
-					}
-			}
-			.frame(height: 14)
-
-			GeometryReader { proxy in
-				ZStack(alignment: .leading) {
-					let width = fillWidth(in: proxy.size.width)
-
-					Capsule()
-						.fill(trackColor)
-						.overlay {
-							Capsule()
-								.fill(trackInsetStyle)
-								.padding(.vertical, 0.8)
-								.allowsHitTesting(false)
-						}
-					Capsule()
-						.fill(fillStyle)
-						.frame(width: width)
-						.clipShape(Capsule())
-						.animation(PanelMotion.state, value: remainingPercent)
-						.shadow(
-							color: color.opacity(colorScheme == .dark ? 0.09 : 0.07),
-							radius: colorScheme == .dark ? 1.2 : 1,
-							x: 0,
-							y: 0
-						)
-					Capsule()
-						.strokeBorder(trackEdgeColor, lineWidth: 0.24)
-						.allowsHitTesting(false)
-				}
-			}
-			.frame(height: 3.2)
-		}
-		.lineLimit(1)
-		.frame(height: 22)
-		.frame(maxWidth: .infinity, alignment: .leading)
-		.accessibilityLabel(accessibilityText)
-	}
-
-	private var remainingText: String {
-		guard let remainingPercent else {
-			return "-"
-		}
-
-		return "\(remainingPercent)% left"
-	}
-
-	private var dailyAverageText: String? {
-		guard let dailyAveragePercent else {
-			return nil
-		}
-		let formatted = formatDailyUsageRate(dailyAveragePercent)
-
-		return formatted == "-" ? nil : formatted
-	}
-
-	private var accessibilityText: String {
-		let average = dailyAverageText.map { ", daily average \($0)" } ?? ""
-		return "\(label) remaining \(remainingText)\(average), \(resetDisplay.accessibility)"
-	}
-
-	private var progress: CGFloat {
-		guard let remainingPercent else {
-			return 0
-		}
-
-		return CGFloat(max(0, min(100, remainingPercent))) / 100
-	}
-
-	private func fillWidth(in width: CGFloat) -> CGFloat {
-		guard remainingPercent != nil else {
-			return 0
-		}
-
-		return max(4, width * progress)
-	}
-
-	private var color: Color {
-		switch tone {
-		case .codexActive: return PanelPalette.codexAccent(colorScheme)
-		case .ready: return PanelPalette.capacityAccent(colorScheme)
-		case .selected: return PanelPalette.routeAccent(colorScheme)
-		case .warning: return PanelPalette.warning(colorScheme)
-		case .danger: return PanelPalette.destructive(colorScheme)
-		case .neutral: return PanelPalette.secondaryText(colorScheme)
-		}
-	}
-
-	private var valueColor: Color {
-		switch tone {
-		case .warning, .danger:
-			return color.opacity(colorScheme == .dark ? 0.95 : 0.78)
-		default:
-			return PanelPalette.primaryText(colorScheme).opacity(colorScheme == .dark ? 0.9 : 0.84)
-		}
-	}
-
-	private var resetDisplay: UsageResetDisplay {
-		UsageResetDisplay.make(resetAtUnixEpoch: resetAtUnixEpoch, now: currentTime)
-	}
-
-	private var trackColor: Color {
-		PanelPalette.progressTrack(colorScheme)
-	}
-
-	private var trackEdgeColor: Color {
-		PanelPalette.progressEdge(colorScheme)
-	}
-
-	private var fillStyle: LinearGradient {
-		LinearGradient(
-			colors: [
-				color.opacity(colorScheme == .dark ? 0.78 : 0.68),
-				color.opacity(colorScheme == .dark ? 0.62 : 0.52),
-			],
-			startPoint: .leading,
-			endPoint: .trailing
-		)
-	}
-
-	private var trackInsetStyle: LinearGradient {
-		LinearGradient(
-			colors: [
-				Color.white.opacity(colorScheme == .dark ? 0.022 : 0.05),
-				Color.white.opacity(0),
-				Color.black.opacity(colorScheme == .dark ? 0.035 : 0.018),
-			],
-			startPoint: .top,
-			endPoint: .bottom
-		)
-	}
-}
-
-private struct UsageGlassTrackView: View {
-	let progress: CGFloat
-	let tint: Color
-	let markers: [CGFloat]
-	let alertMarker: CGFloat?
-	@Environment(\.colorScheme) private var colorScheme
-
-	var body: some View {
-		GeometryReader { proxy in
-			let trackWidth = proxy.size.width
-			let boundedProgress = max(0, min(1, progress))
-			let fillWidth = max(boundedProgress > 0 ? 5 : 0, trackWidth * boundedProgress)
-
-			ZStack(alignment: .leading) {
-				Capsule()
-					.fill(trackFill)
-					.overlay {
-						Capsule()
-							.strokeBorder(PanelPalette.progressEdge(colorScheme), lineWidth: 0.4)
-					}
-
-				Capsule()
-					.fill(fillFill)
-					.frame(width: fillWidth)
-					.shadow(color: tint.opacity(colorScheme == .dark ? 0.18 : 0.12), radius: 2, x: 0, y: 0)
-					.animation(PanelMotion.state, value: progress)
-
-				ForEach(markers, id: \.self) { marker in
-					Rectangle()
-						.fill(markerColor)
-						.frame(width: 1.2)
-						.padding(.vertical, 0.8)
-						.offset(x: markerOffset(marker, in: trackWidth))
-				}
-
-				if let alertMarker {
-					Rectangle()
-						.fill(PanelPalette.destructive(colorScheme))
-						.frame(width: 2)
-						.padding(.vertical, 0.2)
-						.offset(x: markerOffset(alertMarker, in: trackWidth))
-				}
-			}
-		}
-		.accessibilityHidden(true)
-	}
-
-	private var trackFill: LinearGradient {
-		LinearGradient(
-			colors: [
-				PanelPalette.progressTrack(colorScheme).opacity(0.92),
-				PanelPalette.progressTrack(colorScheme).opacity(colorScheme == .dark ? 0.72 : 0.82),
-			],
-			startPoint: .top,
-			endPoint: .bottom
-		)
-	}
-
-	private var fillFill: LinearGradient {
-		LinearGradient(
-			colors: [
-				tint.opacity(colorScheme == .dark ? 0.9 : 0.78),
-				tint.opacity(colorScheme == .dark ? 0.72 : 0.64),
-			],
-			startPoint: .leading,
-			endPoint: .trailing
-		)
-	}
-
-	private var markerColor: Color {
-		colorScheme == .dark
-			? Color.white.opacity(0.48)
-			: Color.white.opacity(0.76)
-	}
-
-	private func markerOffset(_ marker: CGFloat, in width: CGFloat) -> CGFloat {
-		max(0, min(width - 1.2, width * max(0, min(1, marker))))
-	}
-}
-
-struct UsageResetDisplay {
-	let short: String
-	let date: String
-	let accessibility: String
-
-	static func make(resetAtUnixEpoch: Int?, now: Date = Date()) -> UsageResetDisplay {
-		guard let seconds = resetAtUnixEpoch, seconds > 0 else {
-			return UsageResetDisplay(
-				short: "-",
-				date: "",
-				accessibility: "reset unavailable"
-			)
-		}
-
-		let resetAt = Date(timeIntervalSince1970: TimeInterval(seconds))
-		guard resetAt.timeIntervalSince1970.isFinite else {
-			return UsageResetDisplay(
-				short: "unknown",
-				date: "",
-				accessibility: "remaining unknown"
-			)
-		}
-
-		let distanceSeconds = Int(floor(resetAt.timeIntervalSince(now)))
-		if distanceSeconds <= 0 {
-			let date = formatResetDate(resetAt, now: now)
-			return UsageResetDisplay(
-				short: "0m",
-				date: date,
-				accessibility: "reset at \(date), reset due now"
-			)
-		}
-
-		let short = formatResetDuration(distanceSeconds)
-		let date = formatResetDate(resetAt, now: now)
-		return UsageResetDisplay(
-			short: short,
-			date: date,
-			accessibility: "reset at \(date), resets in \(short)"
-		)
-	}
-
-	private static func formatResetDuration(_ seconds: Int) -> String {
-		let value = max(0, seconds)
-		if value < 60 {
-			return "<1m"
-		}
-
-		let days = value / 86_400
-		let hours = (value % 86_400) / 3_600
-		let minutes = (value % 3_600) / 60
-
-		if days > 0 {
-			return hours > 0 ? "\(days)d \(hours)h" : "\(days)d"
-		}
-
-		if hours > 0 {
-			return "\(hours)h \(minutes)m"
-		}
-
-		return "\(minutes)m"
-	}
-
-	private static func formatResetDate(_ date: Date, now: Date) -> String {
-		let formatter = DateFormatter()
-		formatter.locale = Locale(identifier: "en_US_POSIX")
-		let calendar = Calendar(identifier: .gregorian)
-		formatter.dateFormat = calendar.component(.year, from: date) == calendar.component(.year, from: now)
-			? "MMM d HH:mm"
-			: "MMM d yyyy HH:mm"
-		return formatter.string(from: date)
-	}
-}
-
-struct NoticeView: View {
-	let text: String
-	@Environment(\.colorScheme) private var colorScheme
-
-	var body: some View {
-		HStack(alignment: .top, spacing: 7) {
-			Image(systemName: "exclamationmark.triangle")
-				.foregroundStyle(PanelPalette.warning(colorScheme))
-			Text(text)
-				.font(PanelFont.notice)
-				.foregroundStyle(PanelPalette.secondaryText(colorScheme))
-				.fixedSize(horizontal: false, vertical: true)
-		}
-		.padding(8)
-		.modernGlassSurface(
-			cornerRadius: 9,
-			depth: .section
-		)
 	}
 }
 
@@ -3022,22 +2130,6 @@ struct OperatorLaneReadoutDivider: View {
 	}
 }
 
-private struct PanelMetricIconView: View {
-	let symbol: String
-	let tint: Color
-
-	var body: some View {
-		Image(systemName: symbol)
-			.font(PanelFont.summaryIcon)
-			.symbolRenderingMode(.monochrome)
-			.foregroundStyle(tint)
-			.frame(width: 12, height: 12)
-			.alignmentGuide(.firstTextBaseline) { dimensions in
-				dimensions[VerticalAlignment.center] + 3.85
-			}
-	}
-}
-
 struct SummaryTileView: View {
 	let title: String
 	let value: String
@@ -3216,244 +2308,6 @@ private enum AccountPrivacy {
 	static let visibleValue = "visible"
 }
 
-enum AccountDisplay {
-	static let randomNames = [
-		"Alex",
-		"Avery",
-		"Bailey",
-		"Blake",
-		"Casey",
-		"Charlie",
-		"Clara",
-		"Dana",
-		"Drew",
-		"Eden",
-		"Elliot",
-		"Emery",
-		"Evan",
-		"Finley",
-		"Harper",
-		"Hayden",
-		"Iris",
-		"Jamie",
-		"Jordan",
-		"Kai",
-		"Kendall",
-		"Lane",
-		"Liam",
-		"Logan",
-		"Mason",
-		"Maya",
-		"Mia",
-		"Morgan",
-		"Noah",
-		"Nora",
-		"Owen",
-		"Paige",
-		"Parker",
-		"Quinn",
-		"Reese",
-		"Remy",
-		"Riley",
-		"Rowan",
-		"Sage",
-		"Sasha",
-		"Sidney",
-		"Taylor",
-		"Theo",
-		"Val",
-	]
-
-	static func alias(for account: CodexAccount) -> String {
-		randomNames[preferredNameIndex(for: account)]
-	}
-
-	static func aliases(for accounts: [CodexAccount]) -> [String: String] {
-		var usedNames = Set<String>()
-		var aliases = [String: String]()
-		let orderedAccounts = accounts.sorted { left, right in
-			aliasSortKey(for: left) < aliasSortKey(for: right)
-		}
-
-		for account in orderedAccounts {
-			let alias = uniqueAlias(startingAt: preferredNameIndex(for: account), usedNames: usedNames)
-			usedNames.insert(alias)
-			aliases[account.id] = alias
-		}
-
-		return aliases
-	}
-
-	static func alias(forIdentity identity: String) -> String {
-		let seed = identity.trimmingCharacters(in: .whitespacesAndNewlines)
-		let hash = identityHash(seed.isEmpty ? "account" : seed)
-		let index = Int(hash % UInt32(randomNames.count))
-
-		return randomNames[index]
-	}
-
-	static func compactEmail(_ email: String) -> String {
-		let text = email.trimmingCharacters(in: .whitespacesAndNewlines)
-		guard let atIndex = text.firstIndex(of: "@"), atIndex > text.startIndex else {
-			return compactIdentity(text)
-		}
-
-		let local = String(text[..<atIndex])
-		let domain = String(text[atIndex...])
-		if local.count <= 6 {
-			return "\(local)\(domain)"
-		}
-
-		return "\(local.prefix(3))...\(compactLocalSuffix(local))\(domain)"
-	}
-
-	static func compactIdentity(_ value: String) -> String {
-		let text = trimLeadingEllipsis(value)
-		if text.isEmpty || text == "unknown" {
-			return text
-		}
-
-		let edgeLength = max(3, min(6, text.count / 2))
-		return "\(text.prefix(edgeLength))...\(text.suffix(edgeLength))"
-	}
-
-	private static func trimLeadingEllipsis(_ value: String) -> String {
-		let text = value.trimmingCharacters(in: .whitespacesAndNewlines)
-		if text.hasPrefix("..."), text.dropFirst(3).contains("...") == false {
-			return String(text.dropFirst(3))
-		}
-
-		return text
-	}
-
-	private static func compactLocalSuffix(_ local: String) -> String {
-		if let separator = local.lastIndex(of: ".") {
-			let segment = String(local[local.index(after: separator)...])
-			if (2...4).contains(segment.count), segment.allSatisfy(\.isLetter) {
-				return segment
-			}
-		}
-
-		return String(local.suffix(3))
-	}
-
-	private static func identityHash(_ value: String) -> UInt32 {
-		var hash: UInt32 = 2_166_136_261
-		for unit in value.utf16 {
-			hash ^= UInt32(unit)
-			hash = hash &* 16_777_619
-		}
-
-		return hash
-	}
-
-	private static func aliasSortKey(for account: CodexAccount) -> String {
-		if let key = account.randomNameKey?.trimmingCharacters(in: .whitespacesAndNewlines),
-			key.isEmpty == false
-		{
-			return key
-		}
-
-		return account.randomNameSeed
-	}
-
-	private static func preferredNameIndex(for account: CodexAccount) -> Int {
-		if let randomName = account.randomName?.trimmingCharacters(in: .whitespacesAndNewlines),
-			let index = randomNames.firstIndex(of: randomName)
-		{
-			return index
-		}
-
-		let hash = randomNameHash(for: account)
-		let offset = normalizedOffset(account.randomNameOffset ?? 0)
-
-		return (Int(hash % UInt32(randomNames.count)) + offset) % randomNames.count
-	}
-
-	private static func randomNameHash(for account: CodexAccount) -> UInt32 {
-		if let key = account.randomNameKey?.trimmingCharacters(in: .whitespacesAndNewlines),
-			key.isEmpty == false,
-			let hash = UInt32(key, radix: 16)
-		{
-			return hash
-		}
-
-		return identityHash(account.randomNameSeed)
-	}
-
-	private static func normalizedOffset(_ offset: Int) -> Int {
-		((offset % randomNames.count) + randomNames.count) % randomNames.count
-	}
-
-	private static func uniqueAlias(startingAt startIndex: Int, usedNames: Set<String>) -> String {
-		for probe in 0..<randomNames.count {
-			let name = randomNames[(startIndex + probe) % randomNames.count]
-			if usedNames.contains(name) == false {
-				return name
-			}
-		}
-
-		let baseName = randomNames[startIndex % randomNames.count]
-		var suffix = 2
-		while true {
-			let name = "\(baseName) \(suffix)"
-			if usedNames.contains(name) == false {
-				return name
-			}
-			suffix += 1
-		}
-	}
-}
-
-private func formatUsagePercent(_ value: Double) -> String {
-	guard value.isFinite else {
-		return "-"
-	}
-
-	let rounded = value.rounded()
-	if abs(value - rounded) < 0.05 {
-		return "\(Int(rounded))%"
-	}
-
-	return String(format: "%.1f%%", value)
-}
-
-private func formatDailyUsageRate(_ value: Double) -> String {
-	let percent = formatUsagePercent(value)
-	guard percent != "-" else {
-		return "-"
-	}
-
-	return "\(percent)/d"
-}
-
-private func compactUsageDate(_ value: String) -> String {
-	let formatter = DateFormatter()
-	formatter.locale = Locale(identifier: "en_US_POSIX")
-	formatter.dateFormat = "yyyy-MM-dd"
-	guard let date = formatter.date(from: value) else {
-		return value
-	}
-
-	formatter.dateFormat = "MMM d"
-	return formatter.string(from: date)
-}
-
-private func formatPercentagePointDelta(_ value: Double) -> String {
-	guard value.isFinite else {
-		return "-"
-	}
-
-	let absValue = abs(value)
-	let sign = value > 0.05 ? "+" : (value < -0.05 ? "-" : "")
-	let rounded = absValue.rounded()
-	if abs(absValue - rounded) < 0.05 {
-		return "\(sign)\(Int(rounded))pp"
-	}
-
-	return String(format: "%@%.1fpp", sign, absValue)
-}
-
 private func rawPanelToken(_ value: String) -> String {
 	value.trimmingCharacters(in: .whitespacesAndNewlines)
 }
@@ -3479,70 +2333,6 @@ private func formatLaneDuration(_ seconds: Int?) -> String? {
 	}
 
 	return "\(minutes)m"
-}
-
-private func formatActivityDuration(_ seconds: Int?) -> String? {
-	guard let seconds else {
-		return nil
-	}
-
-	let value = max(0, seconds)
-	if value < 60 {
-		return "\(value)s"
-	}
-
-	let hours = value / 3_600
-	let minutes = (value % 3_600) / 60
-	let remainderSeconds = value % 60
-	if hours > 0 {
-		return minutes > 0 ? "\(hours)h \(minutes)m" : "\(hours)h"
-	}
-	if minutes > 0 {
-		return remainderSeconds > 0 ? "\(minutes)m \(remainderSeconds)s" : "\(minutes)m"
-	}
-
-	return "\(remainderSeconds)s"
-}
-
-private func formatCompactCount(_ value: Int) -> String {
-	let absoluteValue = abs(Double(value))
-	let sign = value < 0 ? "-" : ""
-
-	if absoluteValue >= 1_000_000_000 {
-		return "\(sign)\(formatCompactDecimal(absoluteValue / 1_000_000_000))B"
-	}
-	if absoluteValue >= 1_000_000 {
-		return "\(sign)\(formatCompactDecimal(absoluteValue / 1_000_000))M"
-	}
-	if absoluteValue >= 1_000 {
-		return "\(sign)\(formatCompactDecimal(absoluteValue / 1_000))K"
-	}
-
-	return "\(value)"
-}
-
-private func formatCompactBytes(_ value: Int) -> String {
-	let absoluteValue = max(0, Double(value))
-	if absoluteValue >= 1_073_741_824 {
-		return "\(formatCompactDecimal(absoluteValue / 1_073_741_824))GiB"
-	}
-	if absoluteValue >= 1_048_576 {
-		return "\(formatCompactDecimal(absoluteValue / 1_048_576))MiB"
-	}
-	if absoluteValue >= 1_024 {
-		return "\(formatCompactDecimal(absoluteValue / 1_024))KiB"
-	}
-
-	return "\(max(0, value))B"
-}
-
-private func formatCompactDecimal(_ value: Double) -> String {
-	let rounded = (value * 10).rounded() / 10
-	if rounded >= 10 || rounded.rounded() == rounded {
-		return String(format: "%.0f", rounded)
-	}
-
-	return String(format: "%.1f", rounded)
 }
 
 private func formatPanelTimestamp(_ value: String?) -> String? {
@@ -3792,105 +2582,5 @@ struct ModernGlassSurfaceModifier: ViewModifier {
 		case .control:
 			return 1
 		}
-	}
-}
-
-private extension CodexAccount {
-	var randomNameSeed: String {
-		if accountFingerprint.isEmpty == false {
-			return accountFingerprint
-		}
-		if let email, email.isEmpty == false {
-			return email
-		}
-		if let planType, planType.isEmpty == false {
-			return planType
-		}
-
-		return "account"
-	}
-
-	func panelDisplayName(emailsHidden: Bool) -> String {
-		if emailsHidden {
-			if let randomName = randomName?.trimmingCharacters(in: .whitespacesAndNewlines),
-				randomName.isEmpty == false
-			{
-				return randomName
-			}
-
-			return AccountDisplay.alias(for: self)
-		}
-
-		return AccountDisplay.compactEmail(displayName)
-	}
-
-	func statusDisplayColor(colorScheme: ColorScheme) -> Color {
-		switch statusTone {
-		case .codexActive:
-			return PanelPalette.codexAccent(colorScheme)
-		case .ready:
-			return PanelPalette.secondaryText(colorScheme)
-		case .selected:
-			return PanelPalette.routeAccent(colorScheme)
-		case .warning:
-			return PanelPalette.warning(colorScheme)
-		case .danger:
-			return PanelPalette.destructive(colorScheme)
-		case .neutral:
-			return PanelPalette.secondaryText(colorScheme)
-		}
-	}
-
-	var hasPrimaryUsageData: Bool {
-		primaryRemainingPercent != nil || primaryWindowSeconds != nil || primaryResetsAtUnixEpoch != nil
-	}
-
-	var hasSecondaryUsageData: Bool {
-		secondaryRemainingPercent != nil || secondaryWindowSeconds != nil || secondaryResetsAtUnixEpoch != nil
-	}
-
-	var hasUsageSummary: Bool {
-		hasUsageWindowSummary || hasProfileSummary
-	}
-
-	var hasUsageWindowSummary: Bool {
-		hasPrimaryUsageData || hasSecondaryUsageData
-	}
-
-	var recentUsageRecords: [AccountUsageRecord] {
-		usageRecords ?? []
-	}
-
-	func sevenDayAveragePercent(forWindowSeconds seconds: Int?) -> Double? {
-		guard seconds == 604_800 else {
-			return nil
-		}
-
-		return sevenDayDailyAveragePercent
-	}
-
-	var compactHealthLabel: String? {
-		if isUsageLimited {
-			return compactLimitStatusToken
-		}
-
-		if let token = recoveryAction?.trimmingCharacters(in: .whitespacesAndNewlines),
-			token.isEmpty == false
-		{
-			return token
-		}
-
-		let label = status.trimmingCharacters(in: .whitespacesAndNewlines)
-		return label.isEmpty || label == "available" ? nil : label
-	}
-
-	private var compactLimitStatusToken: String {
-		let reached = rateLimitReachedType?.trimmingCharacters(in: .whitespacesAndNewlines)
-		if let reached, reached.isEmpty == false, reached != "none" {
-			return reached
-		}
-
-		let token = status.trimmingCharacters(in: .whitespacesAndNewlines)
-		return token.isEmpty || token == "available" ? "usage_limited" : token
 	}
 }
